backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.0
httpx==0.25.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
faiss-cpu==1.7.4
numpy==1.24.3
gunicorn==21.2.0

